+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 3.0.0-rc.1
 
 * Feature: Add support for GuardSquare's Proguard (#263)
 * Feature: Add support for GuardSquare's Dexguard (#267)